# api/public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Benevolus",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/campaigns/{campaign}/updates": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "List campaign updates",
                "operationId": "listCampaignUpdates",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign updates retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignUpdateResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create campaign update",
                "operationId": "createCampaignUpdate",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCampaignUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campaign update created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignUpdateResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updates/{update}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaign update",
                "operationId": "getCampaignUpdate",
                "parameters": [
                    {
                        "name": "update",
                        "in": "path",
                        "description": "Update ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign update retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignUpdateResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Delete campaign update",
                "operationId": "deleteCampaignUpdate",
                "parameters": [
                    {
                        "name": "update",
                        "in": "path",
                        "description": "Update ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Campaign update deleted successfully"
                    }
                }
            }
        },
        "/api/campaigns/{campaign}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "List campaign comments",
                "operationId": "listCampaignComments",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comments retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CommentResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a comment",
                "operationId": "createComment",
                "parameters": [
                    {
                        "name": "campaign",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comments/{comment}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update comment",
                "operationId": "updateComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete comment",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Comment deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCampaignUpdateRequest": {
                "required": [
                    "title",
                    "content",
                    "visibleToDonorsOnly"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "content": {
                        "type": "string"
                    },
                    "visibleToDonorsOnly": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "CommentStoreRequest": {
                "required": [
                    "content",
                    "is_anonymous"
                ],
                "properties": {
                    "content": {
                        "type": "string"
                    },
                    "is_anonymous": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "CommentUpdateRequest": {
                "properties": {
                    "content": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CampaignUpdateResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "campaignId": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "visibleToDonorsOnly": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CommentResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "campaignId": {
                        "type": "integer"
                    },
                    "content": {
                        "type": "string"
                    },
                    "isAnonymous": {
                        "type": "boolean"
                    },
                    "likes": {
                        "type": "integer"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true,
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "street": {
                                "type": "string",
                                "nullable": true
                            },
                            "number": {
                                "type": "string",
                                "nullable": true
                            },
                            "city": {
                                "type": "string",
                                "nullable": true
                            },
                            "state": {
                                "type": "string",
                                "nullable": true
                            },
                            "zipcode": {
                                "type": "string",
                                "nullable": true
                            },
                            "country": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}